inventory_management/requirements.txt
asgiref==3.8.1
coverage==7.6.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
mysqlclient==2.2.4
packaging==24.1
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.3.1
pytest-cov==5.0.0
pytest-django==4.8.0
python-dotenv==1.0.1
sqlparse==0.5.1
